+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 See [DocuSign Support Center](https://support.docusign.com/en/releasenotes/) for Product Release Notes.
 
+## [6.2.0] - eSignature API v2.1-21.1.02.00 - 2021-06-03
+### Fixed
+- Issue - EnvelopeFormData::get results in exception (DCM-5835)
+- Issue - [`#138`](https://github.com/docusign/docusign-esign-php-client/issues/138):  Tabs class refers to `array` instead of `ModelList` (DCM-3768)
+
 ## [6.2.0-rc] - eSignature API v2.1-21.1.02.00 - 2021-05-20
 ### Breaking
 - Removed methods `getAccountSettingsExport`,`getSealProviders` from Accounts.
@@ -16,7 +21,7 @@ See [DocuSign Support Center](https://support.docusign.com/en/releasenotes/) for
 - Added support for version v2.1-21.1.02.00 of the DocuSign eSignature API.
 - Updated the SDK release version.
 
-## [6.1.0] - eSignature API v2.1-21.1.01.03 - 2021-04-29
+## [6.1.0] - eSignature API v2.1-21.1.01.03 - 2021-04-23
 ### Added
 - Added new methods `deleteCustomFieldsV2`, `deletePageInfoV2`, `getApplianceEnvelopeInfo`, `getTemplateInfo` to envelopes.
 - Added new method `deleteConnectSecret` to connect.